+++ c/frame/py_frame.cc
@@ -8,6 +8,7 @@
 #include "frame/py_frame.h"
 #include <iostream>
 #include "python/int.h"
+#include "python/tuple.h"
 
 namespace py {
 
@@ -16,13 +17,12 @@ namespace py {
 // Declare Frame's API
 //------------------------------------------------------------------------------
 
-NoArgs Frame::Type::args_bang;
 PKArgs Frame::Type::args___init__(1, 0, 3, false, false,
                                   {"src", "names", "stypes", "stype"});
 
 
 const char* Frame::Type::classname() {
-  return "datatable.Frame";
+  return "datatable.core.Frame";
 }
 
 const char* Frame::Type::classdoc() {
@@ -34,22 +34,39 @@ const char* Frame::Type::classdoc() {
     "Internally the data is stored as C primitives, and processed using\n"
     "multithreaded native C++ code.\n"
     "\n"
-    "This is a primary data structure for datatable module.\n";
+    "This is a primary data structure for the `datatable` module.\n";
 }
 
-void Frame::Type::init_getsetters(GetSetters& gs) {
-  gs.add<&Frame::get_ncols>("ncols");
-  gs.add<&Frame::get_nrows>("nrows");
+void Frame::Type::init_getsetters(GetSetters& gs)
+{
+  gs.add<&Frame::get_ncols>("ncols",
+    "Number of columns in the Frame\n");
+  gs.add<&Frame::get_nrows, &Frame::set_nrows>("nrows",
+    "Number of rows in the Frame.\n"
+    "\n"
+    "Assigning to this property will change the height of the Frame,\n"
+    "either by truncating if the new number of rows is smaller than the\n"
+    "current, or filling with NAs if the new number of rows is greater.\n"
+    "\n"
+    "Increasing the number of rows of a keyed Frame is not allowed.\n");
+  gs.add<&Frame::get_shape>("shape",
+    "Tuple with (nrows, ncols) dimensions of the Frame\n");
+  gs.add<&Frame::get_key>("key");
+  gs.add<&Frame::get_internal>("internal", "[DEPRECATED]");
+  gs.add<&Frame::get_internal, &Frame::set_internal>("_dt");
 }
 
-void Frame::Type::init_methods(Methods& mm) {
-  mm.add<&Frame::bang, args_bang>("bang");
+void Frame::Type::init_methods(Methods&) {
+  // mm.add<&Frame::bang, args_bang>("bang");
 }
 
 
 
 void Frame::m__dealloc__() {
-  std::cout << "In Frame::dealloc (dt=" << dt << ")\n";
+  Py_XDECREF(core_dt);
+  // `dt` is already managed by `core_dt`.
+  // delete dt;
+  dt = nullptr;
 }
 
 void Frame::m__get_buffer__(Py_buffer* , int ) const {
@@ -58,17 +75,54 @@ void Frame::m__get_buffer__(Py_buffer* , int ) const {
 void Frame::m__release_buffer__(Py_buffer*) const {
 }
 
+
+//------------------------------------------------------------------------------
+// Getters / setters
+//------------------------------------------------------------------------------
+
 oobj Frame::get_ncols() const {
-  return oobj(py::oInt(11));
+  return py::oInt(dt->ncols);
 }
+
 
 oobj Frame::get_nrows() const {
-  return oobj(py::oInt(47));
+  return py::oInt(dt->nrows);
 }
 
-oobj Frame::bang(NoArgs&) {
-  std::cout << "Yay, Frame::bang()!\n";
-  return None();
+void Frame::set_nrows(obj nr) {
+  if (!nr.is_int()) {
+    throw TypeError() << "Number of rows must be an integer, not "
+        << nr.typeobj();
+  }
+  int64_t new_nrows = nr.to_int64_strict();
+  if (new_nrows < 0) {
+    throw ValueError() << "Number of rows cannot be negative";
+  }
+  dt->resize_rows(new_nrows);
+}
+
+
+oobj Frame::get_shape() const {
+  py::otuple shape(2);
+  shape.set(0, get_nrows());
+  shape.set(1, get_ncols());
+  return shape;
+}
+
+oobj Frame::get_key() const {
+  py::otuple key(dt->nkeys);
+  // Fill in the keys...
+  return key;
+}
+
+oobj Frame::get_internal() const {
+  return oobj(core_dt);
+}
+
+void Frame::set_internal(obj _dt) {
+  m__dealloc__();
+  dt = _dt.to_frame();
+  core_dt = static_cast<pydatatable::obj*>(_dt.to_pyobject_newref());
 }