+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 #### Changed
 - `names` argument in `Frame()` constructor can no longer be a string --
   use a list or tuple of strings instead.
+- `Frame.resize()` was removed -- same functionality is available via
+  assigning to `Frame.nrows`.
 
 #### Fixed
 - bug in dt.cbind() where the first Frame in the list was ignored.